+++ src/main/java/com/github/mavogel/ilias/state/states/permissions/GrantFileUploadToGroupMembers.java
@@ -1,4 +1,4 @@
-/*
+package com.github.mavogel.ilias.state.states.permissions;/*
  *  The MIT License (MIT)
  *
  *  Copyright (c) 2016 Manuel Vogel
@@ -23,43 +23,23 @@
  *
  *  https://opensource.org/licenses/MIT
  */
-package com.github.mavogel.ilias.state.states.action;
 
 import com.github.mavogel.client.ILIASSoapWebservicePortType;
 import com.github.mavogel.ilias.model.IliasNode;
 import com.github.mavogel.ilias.model.UserDataIds;
 import com.github.mavogel.ilias.state.ChangeAction;
 import com.github.mavogel.ilias.utils.IOUtils;
-import com.github.mavogel.ilias.utils.IliasUtils;
-import org.apache.log4j.Logger;
-import org.jdom.JDOMException;
 
-import java.io.IOException;
 import java.util.List;
 
 /**
- * Represents the action for removing the uploaded materials.
- *
- * Created by mavogel on 9/9/16.
+ * Created by mavogel on 9/20/16.
  */
-public class RemoveUploadedMaterialsChange implements ChangeAction {
-
-    private static Logger LOG = Logger.getLogger(RemoveUploadedMaterialsChange.class);
+public class GrantFileUploadToGroupMembers implements ChangeAction {
 
     @Override
-    public String performAction(final ILIASSoapWebservicePortType endpoint, final UserDataIds userDataIds,
-                                final List<IliasNode> nodes) {
-        LOG.info("Removing uploaded materials from groups");
-        confirm();
-
-        final String sid = userDataIds.getSid();
-        final int userId = userDataIds.getUserId();
-        try {
-            IliasUtils.deleteObjects(endpoint, sid, IliasUtils.retrieveFileRefIdsFromGroups(endpoint, sid, userId, nodes));
-        } catch (IOException | JDOMException e) {
-            LOG.error("Error creating xml parser: " + e.getMessage());
-        }
-        return "";
+    public String performAction(final ILIASSoapWebservicePortType endpoint, final UserDataIds userDataIds, final List<IliasNode> nodes) {
+        return null;
     }
 
     @Override